Make types global
